src/containers/docker/websocket-app/backend/requirements.txt
websockets
asyncio
fastapi
fastapi-socketio
python-socketio
uvicorn
loguru
sqlalchemy
psycopg2-binary
